+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/screen/swap/info/SwapInfoView.kt
@@ -30,8 +30,13 @@ fun SwapInfoView(state: SwapInfoState) {
         onBack = state.onBack,
         primaryButton =
             ButtonState(
-                text = stringRes(co.electriccoin.zcash.ui.design.R.string.general_ok),
+                text = stringRes(co.electriccoin.zcash.ui.design.R.string.general_dismiss),
                 onClick = state.onBack,
+            ),
+        secondaryButton =
+            ButtonState(
+                text = stringRes(R.string.swap_info_learn_more),
+                onClick = state.onLearnMoreClick,
             ),
     ) {
         Row(
@@ -61,4 +66,4 @@ fun SwapInfoView(state: SwapInfoState) {
 
 @PreviewScreens
 @Composable
-private fun Preview() = ZcashTheme { SwapInfoView(state = SwapInfoState(onBack = {})) }
+private fun Preview() = ZcashTheme { SwapInfoView(state = SwapInfoState(onBack = {}, onLearnMoreClick = {})) }